03-project/requirements.txt
absl-py==0.10.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
binghamton-camerafp==1.0.0
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
clang==5.0
click==8.0.3
cloudpickle==2.0.0
cryptography==35.0.0
cycler==0.11.0
cytoolz==0.11.2
dask==2021.11.2
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.3
flatbuffers==1.12
flit_core==3.6.0
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.11.1
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
idna==3.1
imagecodecs==2021.11.20
imageio==2.9.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
locket==0.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
multiprocess==0.70.12.2
munkres==1.1.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
noiseprint==1.0.0
notebook==6.4.6
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathos==0.2.8
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
pooch==1.5.2
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==59.4.0
six==1.15.0
statsmodels==0.13.1
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-macos==2.6.0
tensorflow-metal==0.2.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.7.4.3
unicodedata2==13.0.0.post2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.35.1
widgetsnbextension==3.5.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0
